requirements-dev.txt
Pillow==8.0.1
pytest==6.1.2
pytest-mock==3.3.1
pytest-cov==2.10.1
pytest-html==3.0.0